review remidation
