+++ fence/__init__.py
@@ -64,7 +64,6 @@ import fence.blueprints.google
 import fence.blueprints.privacy
 import fence.blueprints.register
 import fence.blueprints.ga4gh
-import fence.blueprints.bucket_info
 
 
 # for some reason the temp dir does not get created properly if we move
@@ -144,10 +143,6 @@ def app_register_blueprints(app):
 
     app.register_blueprint(fence.blueprints.register.blueprint, url_prefix="/register")
     app.register_blueprint(fence.blueprints.ga4gh.blueprint, url_prefix="/ga4gh")
-
-    app.register_blueprint(
-        fence.blueprints.bucket_info.blueprint, url_prefix="/bucket_info"
-    )
 
     fence.blueprints.misc.register_misc(app)
